Refactor block_fn out of map-like logical operators
