+++ python/ray/data/_internal/planner/filter.py
@@ -1,0 +1,23 @@
+from typing import Callable, Iterator
+
+from ray.data.block import Block, BlockAccessor, RowUDF
+from ray.data.context import DatasetContext
+
+
+def generate_filter_fn() -> Callable[[Iterator[Block]], Iterator[Block]]:
+    """Generate function to apply the UDF to each record of blocks,
+    and filter out records that do not satisfy the given predicate.
+    """
+    context = DatasetContext.get_current()
+
+    def fn(blocks: Iterator[Block], row_fn: RowUDF) -> Iterator[Block]:
+        DatasetContext._set_current(context)
+        for block in blocks:
+            block = BlockAccessor.for_block(block)
+            builder = block.builder()
+            for row in block.iter_rows():
+                if row_fn(row):
+                    builder.add(row)
+            return [builder.build()]
+
+    return fn
